Update Calc.java
changing signature

+++ src/main/java/calculator/Calc.java
@@ -2,21 +2,21 @@ package calculator;
 
 public class Calc {
 	
-	public int add(int a, int b) {
+	public double add(int a, int b) {
 		int result = 0;
 		result = a + b;
 		
 		return result;
 	}
 	
-	public int subtract(int a, int b) {
+	public double subtract(int a, int b) {
 		int result = 0;
 		result = a - b;
 		
 		return result;
 	}
 	
-	public int multiply(int a, int b) {
+	public double multiply(int a, int b) {
 		int result = 0;
 		result = a * b;
